# contracts/openapi.yaml
components:
  schemas:
...
    BusinessProfileRequest:
      additionalProperties: false
      properties:
        business_identity:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Business Identity
        commercial_driver_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Commercial Driver Notes
        competition_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Competition Notes
        customer_journey_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Customer Journey Notes
        decision_process_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Decision Process Notes
        economics_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Economics Notes
        events:
          default: []
          items:
            additionalProperties: true
            type: object
          title: Events
          type: array
        facts:
          default: []
          items:
            additionalProperties: true
            type: object
          title: Facts
          type: array
        hypotheses:
          default: []
          items:
            additionalProperties: true
            type: object
          title: Hypotheses
          type: array
        knowledge_gaps:
          default: []
          items:
            additionalProperties: true
            type: object
          title: Knowledge Gaps
          type: array
        kpi:
          anyOf:
          - type: string
          - type: 'null'
          title: Kpi
        kpi_custom_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Kpi Custom Text
        kpi_definition:
          anyOf:
          - type: string
          - type: 'null'
          title: Kpi Definition
        marketing_portfolio:
          default: []
          items:
            additionalProperties: true
            type: object
          title: Marketing Portfolio
          type: array
        markets:
          default: []
          items:
            additionalProperties: true
            type: object
          title: Markets
          type: array
        measurement_objectives:
          default: []
          items:
            additionalProperties: true
            type: object
          title: Measurement Objectives
          type: array
        metadata:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Metadata
        prior_evidence:
          default: []
          items:
            additionalProperties: true
            type: object
          title: Prior Evidence
          type: array
        relationships:
          default: []
          items:
            additionalProperties: true
            type: object
          title: Relationships
          type: array
      title: BusinessProfileRequest
      type: object
...
    ClarificationAnswerRequest:
      additionalProperties: false
      properties:
        answer:
          title: Answer
          type: string
        observed_evidence:
          additionalProperties: true
          default: {}
          title: Observed Evidence
          type: object
      required:
      - answer
      title: ClarificationAnswerRequest
      type: object
    ClarificationCreateRequest:
      additionalProperties: false
      properties:
        coverage_gap_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Coverage Gap Id
        fact_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Fact Id
        question:
          title: Question
          type: string
      required:
      - question
      title: ClarificationCreateRequest
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ProposalCreateRequest:
      additionalProperties: false
      properties:
        observed_evidence:
          additionalProperties: true
          default: {}
          title: Observed Evidence
          type: object
        previous_fact:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Previous Fact
        proposed_fact:
          additionalProperties: true
          default: {}
          title: Proposed Fact
          type: object
      title: ProposalCreateRequest
      type: object
    ProposalDecideRequest:
      additionalProperties: false
      properties:
        accept:
          title: Accept
          type: boolean
      required:
      - accept
      title: ProposalDecideRequest
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    HTTPBearer:
      bearerFormat: JWT
      description: Clerk session token forwarded by the Next.js BFF. prem3-api verifies the session token. Tenant IDs are
        never accepted from the client.
      scheme: bearer
      type: http
info:
  description: Presentation-safe Project, Dataset, upload, Evaluation, catalog, billing, Google connection, import/publish
    governance, Business IQ, and Data Foundation contracts. Clerk session tokens are verified when the identity provider is
    configured. Creating an Evaluation returns 202 Accepted for resource creation only; durable ADK dispatch is not started
    from this HTTP boundary. Tenant identity is never accepted from the client. IMPORT_READY, FOUNDATION_SOURCE_READY, DATA_FOUNDATION_READY,
    MODEL_READY, and PUBLISH_READY are distinct deterministic states.
  summary: PreM3 authenticated product API
  title: prem3-api
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /v1/workspaces/{workspace_id}/business-iq/brief:
    get:
      operationId: getBusinessIntelligenceBrief
      parameters:
      - in: path
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Getbusinessintelligencebrief
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get Brief
      tags:
      - business-iq
    post:
      operationId: regenerateBusinessIntelligenceBrief
      parameters:
      - in: path
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Regeneratebusinessintelligencebrief
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Regenerate Brief
      tags:
      - business-iq
  /v1/workspaces/{workspace_id}/business-iq/clarifications:
    post:
      operationId: createBusinessClarification
      parameters:
      - in: path
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClarificationCreateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Createbusinessclarification
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Create Clarification
      tags:
      - business-iq
  /v1/workspaces/{workspace_id}/business-iq/clarifications/{clarification_id}/answer:
    post:
      operationId: answerBusinessClarification
      parameters:
      - in: path
        name: clarification_id
        required: true
        schema:
          title: Clarification Id
          type: string
      - in: path
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClarificationAnswerRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Answerbusinessclarification
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Answer Clarification
      tags:
      - business-iq
  /v1/workspaces/{workspace_id}/business-iq/profile:
    get:
      operationId: getBusinessProfile
      parameters:
      - in: path
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Getbusinessprofile
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get Profile
      tags:
      - business-iq
    patch:
      operationId: patchBusinessProfile
      parameters:
      - in: path
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessProfileRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Patchbusinessprofile
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Patch Profile
      tags:
      - business-iq
    post:
      operationId: createBusinessProfile
      parameters:
      - in: path
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessProfileRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Createbusinessprofile
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Create Profile
      tags:
      - business-iq
  /v1/workspaces/{workspace_id}/business-iq/profile/snapshots/{snapshot_id}:
    get:
      operationId: getBusinessProfileSnapshot
      parameters:
      - in: path
        name: snapshot_id
        required: true
        schema:
          title: Snapshot Id
          type: string
      - in: path
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Getbusinessprofilesnapshot
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get Snapshot
      tags:
      - business-iq
  /v1/workspaces/{workspace_id}/business-iq/profile/versions:
    get:
      operationId: listBusinessProfileVersions
      parameters:
      - in: path
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Listbusinessprofileversions
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: List Versions
      tags:
      - business-iq
  /v1/workspaces/{workspace_id}/business-iq/proposals:
    get:
      operationId: listBusinessProfileProposals
      parameters:
      - in: path
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Listbusinessprofileproposals
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: List Proposals
      tags:
      - business-iq
    post:
      operationId: createBusinessProfileProposal
      parameters:
      - in: path
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProposalCreateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Createbusinessprofileproposal
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Create Proposal
      tags:
      - business-iq
  /v1/workspaces/{workspace_id}/business-iq/proposals/{proposal_id}/decide:
    post:
      operationId: decideBusinessProfileProposal
      parameters:
      - in: path
        name: proposal_id
        required: true
        schema:
          title: Proposal Id
          type: string
      - in: path
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProposalDecideRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Decidebusinessprofileproposal
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Decide Proposal
      tags:
      - business-iq
...